Move extension handling to new library
