+++ CesiumJsonExtensions/src/ExtensionContext.cpp
@@ -1,8 +1,9 @@
-#include "CesiumJsonReader/ExtensionContext.h"
-#include "CesiumJsonReader/IExtensionJsonHandler.h"
-#include "CesiumJsonReader/JsonObjectJsonHandler.h"
-#include "CesiumJsonReader/JsonReader.h"
+#include "CesiumJsonExtensions/ExtensionContext.h"
+#include "CesiumJsonExtensions/IExtensionJsonHandler.h"
+#include <CesiumJsonReader/JsonObjectJsonHandler.h>
+#include <CesiumJsonReader/JsonReader.h>
 
+using namespace CesiumJsonExtensions;
 using namespace CesiumJsonReader;
 using namespace CesiumUtility;
